Move arch.h to util
Summary: And use decltype instead of boost's type_traits.

Reviewed By: mxw

Differential Revision: D3467213

fbshipit-source-id: 54148614738f6e2428d7aedc92c10b85afb92e2c
