+++ hphp/runtime/vm/jit/arg-group.h
@@ -16,11 +16,11 @@
 #ifndef incl_HPHP_JIT_ARG_GROUP_H
 #define incl_HPHP_JIT_ARG_GROUP_H
 
-#include "hphp/runtime/base/arch.h"
-
 #include "hphp/runtime/vm/jit/containers.h"
 #include "hphp/runtime/vm/jit/reg-alloc.h"
 #include "hphp/runtime/vm/jit/vasm-reg.h"
+
+#include "hphp/util/arch.h"
 
 namespace HPHP { namespace jit {
