+++ hphp/runtime/vm/jit/debug-guards.cpp
@@ -16,7 +16,6 @@
 
 #include "hphp/runtime/vm/jit/debug-guards.h"
 
-#include "hphp/runtime/base/arch.h"
 #include "hphp/runtime/base/request-injection-data.h"
 #include "hphp/runtime/base/thread-info.h"
 #include "hphp/runtime/vm/srckey.h"
@@ -32,6 +31,8 @@
 #include "hphp/runtime/vm/jit/vasm-gen.h"
 #include "hphp/runtime/vm/jit/vasm-instr.h"
 #include "hphp/runtime/vm/jit/vasm-reg.h"
+
+#include "hphp/util/arch.h"
 
 namespace HPHP { namespace jit {
