+++ hphp/runtime/vm/jit/translate-region.cpp
@@ -16,12 +16,12 @@
 
 #include "hphp/runtime/vm/jit/translate-region.h"
 
+#include "hphp/util/arch.h"
 #include "hphp/util/map-walker.h"
 #include "hphp/util/ringbuffer.h"
 #include "hphp/util/timer.h"
 #include "hphp/util/trace.h"
 
-#include "hphp/runtime/base/arch.h"
 #include "hphp/runtime/base/runtime-option.h"
 
 #include "hphp/runtime/vm/bc-pattern.h"
