+++ hphp/runtime/vm/jit/vasm-print.cpp
@@ -18,7 +18,6 @@
 
 #include <type_traits>
 
-#include "hphp/runtime/base/arch.h"
 #include "hphp/runtime/base/stats.h"
 #include "hphp/runtime/vm/jit/mc-generator.h"
 #include "hphp/runtime/vm/jit/print.h"
@@ -29,6 +28,7 @@
 #include "hphp/runtime/vm/jit/vasm-visit.h"
 
 #include "hphp/util/abi-cxx.h"
+#include "hphp/util/arch.h"
 #include "hphp/util/ringbuffer.h"
 #include "hphp/util/stack-trace.h"
