+++ hphp/runtime/vm/jit/vasm-xls.cpp
@@ -16,7 +16,6 @@
 
 #include "hphp/runtime/vm/jit/vasm.h"
 
-#include "hphp/runtime/base/arch.h"
 #include "hphp/runtime/base/stats.h"
 
 #include "hphp/runtime/vm/jit/abi.h"
@@ -32,6 +31,7 @@
 #include "hphp/runtime/vm/jit/vasm-util.h"
 #include "hphp/runtime/vm/jit/vasm-visit.h"
 
+#include "hphp/util/arch.h"
 #include "hphp/util/assertions.h"
 #include "hphp/util/dataflow-worklist.h"
 #include "hphp/util/trace.h"
